finance_agent_team/requirements.txt
python-dotenv
google-genai
agno
